pyproject.toml
[project]
name = "transcribe"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai",
    "pydantic>=2.12.5",
    "send2trash>=2.1.0",
]
